# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: ProfitFlow API
  version: 1.0
  description: Documentação da API ProfitFlow, com módulos de Pessoas, Transações, Financeiro, Relatórios e Autenticação

tags:
  - name: Pessoas
    description: Operações relacionadas a Clientes e Fornecedores
  - name: Transacoes
    description: Operações relacionadas a Receitas e Despesas
  - name: Financeiro
    description: Operações relacionadas a Contas a pagar e a receber
  - name: Relatorios
    description: Operações relacionadas a relatórios financeiros
  - name: Auth
    description: Operações relacionadas a autenticação de usuários

paths:
  # --- Módulo Pessoas ---
  /api/pessoas:
    get:
      summary: Lista todas as pessoas
      tags:
        - Pessoas
      responses:
        '200':
          description: Lista de pessoas retornada com sucesso
    post:
      summary: Cadastra uma nova pessoa
      tags:
        - Pessoas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPessoaDTO'
      responses:
        '201':
          description: Pessoa criada com sucesso
        '400':
          description: Dados inválidos
  /api/pessoas/{id}:
    get:
      summary: Retorna uma pessoa específica
      tags:
        - Pessoas
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pessoa encontrada
        '404':
          description: Pessoa não encontrada
    put:
      summary: Atualiza uma pessoa
      tags:
        - Pessoas
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPessoaDTO'
      responses:
        '200':
          description: Pessoa atualizada
        '400':
          description: Dados inválidos
        '404':
          description: Pessoa não encontrada
    delete:
      summary: Remove uma pessoa
      tags:
        - Pessoas
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Pessoa removida
        '404':
          description: Pessoa não encontrada
  /api/pessoas/search:
    get:
      summary: Busca pessoas por nome, CPF ou CNPJ
      tags:
        - Pessoas
      parameters:
        - in: query
          name: nome
          schema:
            type: string
      responses:
        '200':
          description: Resultado da busca
        '404':
          description: Nenhuma pessoa encontrada
  /api/pessoas/tipo/{tipo}:
    get:
      summary: Lista pessoas de um tipo específico (CLIENTE ou FORNECEDOR)
      tags:
        - Pessoas
      parameters:
        - in: path
          name: tipo
          required: true
          schema:
            type: string
            enum: [CLIENTE, FORNECEDOR]
      responses:
        '200':
          description: Lista filtrada por tipo

  # --- Módulo Transações Financeiras ---
  /api/transacoes:
    get:
      summary: Lista todas as transações
      tags:
        - Transacoes
      responses:
        '200':
          description: Lista de transações retornada
    post:
      summary: Registra uma nova transação
      tags:
        - Transacoes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTransacaoDTO'
      responses:
        '201':
          description: Transação criada
        '400':
          description: Dados inválidos
  /api/transacoes/{id}:
    get:
      summary: Retorna uma transação específica
      tags:
        - Transacoes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transação encontrada
        '404':
          description: Transação não encontrada
    put:
      summary: Atualiza uma transação
      tags:
        - Transacoes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTransacaoDTO'
      responses:
        '200':
          description: Transação atualizada
        '400':
          description: Dados inválidos
        '404':
          description: Transação não encontrada
    delete:
      summary: Remove uma transação
      tags:
        - Transacoes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Transação removida
        '404':
          description: Transação não encontrada
  /api/transacoes/periodo:
    get:
      summary: Filtra transações por período
      tags:
        - Transacoes
      parameters:
        - in: query
          name: inicio
          schema:
            type: string
            format: date
        - in: query
          name: fim
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista filtrada por período

  # --- Módulo Financeiro (Contas) ---
  /api/financeiro:
    get:
      summary: Lista todas as contas
      tags:
        - Financeiro
      responses:
        '200':
          description: Lista de contas retornada
    post:
      summary: Registra nova conta
      tags:
        - Financeiro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestContaDTO'
      responses:
        '201':
          description: Conta criada
        '400':
          description: Dados inválidos
  /api/financeiro/{id}:
    get:
      summary: Retorna uma conta específica
      tags:
        - Financeiro
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Conta encontrada
        '404':
          description: Conta não encontrada
    put:
      summary: Atualiza uma conta
      tags:
        - Financeiro
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestContaDTO'
      responses:
        '200':
          description: Conta atualizada
        '400':
          description: Dados inválidos
        '404':
          description: Conta não encontrada
    delete:
      summary: Remove uma conta
      tags:
        - Financeiro
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Conta removida
        '404':
          description: Conta não encontrada
    patch:
      summary: Marca conta como quitada
      tags:
        - Financeiro
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Conta quitada
        '400':
          description: Dados inválidos
        '404':
          description: Conta não encontrada
  /api/financeiro/tipo/{tipo}:
    get:
      summary: Lista contas por tipo (PAGAR ou RECEBER)
      tags:
        - Financeiro
      parameters:
        - in: path
          name: tipo
          required: true
          schema:
            type: string
            enum: [PAGAR, RECEBER]
      responses:
        '200':
          description: Lista filtrada por tipo
        '404':
          description: Nenhuma conta encontrada

  # --- Módulo Relatórios ---
  /api/relatorios/fluxo-caixa:
    get:
      summary: Retorna fluxo de caixa por período
      tags:
        - Relatorios
      parameters:
        - in: query
          name: inicio
          schema:
            type: string
            format: date
        - in: query
          name: fim
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Relatório gerado
        '204':
          description: Sem dados no período
  /api/relatorios/balanco-mensal:
    get:
      summary: Retorna balanço mensal consolidado
      tags:
        - Relatorios
      parameters:
        - in: query
          name: mes
          schema:
            type: integer
        - in: query
          name: ano
          schema:
            type: integer
      responses:
        '200':
          description: Relatório gerado
        '204':
          description: Sem dados para o mês
  /api/relatorios/situacao-financeira:
    get:
      summary: Retorna resumo geral da saúde financeira
      tags:
        - Relatorios
      responses:
        '200':
          description: Relatório gerado

  # --- Módulo Autenticação ---
  /api/auth/login:
    post:
      summary: Realiza login e gera token JWT
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestLoginDTO'
      responses:
        '200':
          description: Login bem-sucedido
        '401':
          description: Credenciais inválidas
        '400':
          description: Requisição inválida
  /api/auth/register:
    post:
      summary: Registra novo usuário
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestRegisterDTO'
      responses:
        '201':
          description: Usuário registrado
        '400':
          description: Dados inválidos
        '409':
          description: Usuário já existe
  /api/auth/profile:
    get:
      summary: Retorna dados do usuário autenticado
      tags:
        - Auth
      responses:
        '200':
          description: Perfil do usuário
        '401':
          description: Não autenticado
  /api/auth/logout:
    post:
      summary: Invalida token JWT
      tags:
        - Auth
      responses:
        '200':
          description: Logout realizado com sucesso

components:
  schemas:
    RequestPessoaDTO:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        cpfCnpj:
          type: string
        telefone:
          type: string
        endereco:
          type: string
        pessoa:
          type: string
          enum: [CLIENTE, FORNECEDOR]
    RequestTransacaoDTO:
      type: object
      properties:
        valor:
          type: number
        tipo:
          type: string
          enum: [RECEITA, DESPESA]
        categoria:
          type: string
        data:
          type: string
          format: date
        status:
          type: string
    RequestContaDTO:
      type: object
      properties:
        descricao:
          type: string
        valor:
          type: number
        tipo:
          type: string
          enum: [PAGAR, RECEBER]
        vencimento:
          type: string
          format: date
        status:
          type: string
    RequestLoginDTO:
      type: object
      properties:
        email:
          type: string
        senha:
          type: string
    RequestRegisterDTO:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        senha:
          type: string